ondemand: fast start detection via low-power idle notifier
Implement a heuristic to detect interactive usage of the device and
boost the cpu frequency to the max.

A new parameter is provided in the ondemand cpufreq governor and
exported in sysfs:

  /sys/devices/system/cpu/cpufreq/ondemand/fast_start

The fast start detection is implemented using the following heuristic.

If in the last 4 samples the cpu never entered in low-power idle state,
probably the cpu is very busy. In this case we jump to the maximum
frequency immediately.

Signed-off-by: Andrea Righi <andrea@betterlinux.com>